requirements.txt
Flask==3.1.1
knock-subdomains==7.0.2